% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_progression_rate_model}
\alias{fit_progression_rate_model}
\title{Fit a progression rate model to case-carrier data}
\usage{
fit_progression_rate_model(
  input_data,
  type_specific = TRUE,
  location_adjustment = TRUE,
  stat_model = "poisson",
  strain_as_primary_type = FALSE,
  strain_as_secondary_type = FALSE,
  model_description = NULL,
  num_chains = 4,
  num_iter = 10000,
  num_cores = 1,
  adapt_delta_value = 0.8,
  stepsize_value = 1,
  max_treedepth_value = 10
)
}
\arguments{
\item{input_data}{List of lists generated by `process_input_data`}

\item{type_specific}{Boolean specifying whether progression rates vary between types}

\item{location_adjustment}{Boolean specifying whether progression rates vary between locations}

\item{stat_model}{Whether progression to disease is "poisson" (Poisson process) or "negbin" (overdispersed negative binomial distribution)}

\item{strain_as_primary_type}{Whether strain should be used as the primary determinant of progression rate, and the other type used as the secondary determinant}

\item{strain_as_secondary_type}{Whether strain should be used as the secondary determinant of progression rate, and the other type used as the primary determinant}

\item{model_description}{Descriptive title of model}

\item{num_chains}{Number of MCMCs to be run for inference}

\item{num_iter}{Length of MCMCs}

\item{num_cores}{Number of threads used to calculate MCMCs}

\item{adapt_delta_value}{Target average acceptance probability of MCMCs (default = 0.8)}

\item{stepsize_value}{Initial MCMC step size (default = 1)}

\item{max_treedepth_value}{Depth of tree explored by MCMC sampler (default = 10)}
}
\value{
A stanfit object
}
\description{
Fit a progression rate model to case-carrier data
}
