% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_input_data}
\alias{process_input_data}
\title{Generate model input from input data frame}
\usage{
process_input_data(
  input_df,
  type = "type",
  use_strain = FALSE,
  combine_strain = FALSE,
  condense = FALSE
)
}
\arguments{
\item{input_df}{Data frame containing details of case-carrier studies}

\item{type}{Name of column in input data frame used for typing information}

\item{use_strain}{Boolean specifying whether strain information should be used in addition to type information}

\item{combine_strain}{Boolean specifying whether strain information should be combined with type information}

\item{condense}{Boolean specifying whether repeated entries of the same type should be condensed into a single entry}
}
\value{
A list of lists used as an input to stan models
}
\description{
Generate model input from input data frame
}
