% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_input_xlsx}
\alias{process_input_xlsx}
\title{Generate data frame from input spreadsheet}
\usage{
process_input_xlsx(
  fn = "progression_estimation_input.xlsx",
  use_strain = FALSE
)
}
\arguments{
\item{fn}{Name of spreadsheet file containing data}

\item{use_strain}{Boolean variable indicating whether the strain column should also be used for subtyping}
}
\value{
Data frame containing data extracted from spreadsheet
}
\description{
Generate data frame from input spreadsheet
}
