% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressionEstimation-package.R
\docType{package}
\name{progressionEstimation-package}
\alias{progressionEstimation-package}
\alias{progressionEstimation}
\title{The 'progressionEstimation' package.}
\description{
This package uses Bayesian models implemented in stan to estimate the rates at which microbes progress from carriage to disease using case and carrier data.
}
