% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_progression_rate_model}
\alias{fit_progression_rate_model}
\title{Fit a progression rate model to input data}
\usage{
fit_progression_rate_model(
  input_data,
  type_specific = TRUE,
  location_adjustment = TRUE,
  stat_model = "poisson",
  strain_as_primary_type = FALSE,
  strain_as_secondary_type = FALSE,
  num_chains = 4,
  num_iter = 10000,
  adapt_delta = 0.8
)
}
\arguments{
\item{adapt_delta}{}
}
\value{

}
\description{
Fit a progression rate model to input data
}
