% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_input_data}
\alias{process_input_data}
\title{Generate model input from input data frame}
\usage{
process_input_data(input_df, subtype = "categorisation")
}
\arguments{
\item{input_df}{Data frame containing details of case-carrier studies}

\item{subtype}{Name of column in input data frame used for subtyping information}
}
\value{
A list of lists used as an input to stan models
}
\description{
Generate model input from input data frame
}
\examples{
process_input_data(S_pneumoniae_infant_serotype, subtype = "categorisation")
}
